+++ src/Sulu/Bundle/AdminBundle/Resources/public/js/components/header/main.js
@@ -361,6 +361,7 @@ define([], function () {
                         position: 30,
                         items: [
                             {
+                                id: "delete-button",
                                 title: this.sandbox.translate('toolbar.delete'),
                                 callback: function () {
                                     this.sandbox.emit('sulu.header.toolbar.delete');